Specify to use FILE_COMPILE=file while cross compiling
otherwise it tries to use file.exe which is obviously not present on linux.
Note that it does not mitigate the problem of local file needing to match
the build file version.  So those need to be "matched". I will be progressing
submodule in the next commit to match the one I have locally